+++ Engine/Pieces/Piece.cs
@@ -15,7 +15,7 @@ namespace Engine
             this.Position = position;
         }
 
-        public abstract double GetValue();
+        public abstract float GetValue();
         public abstract Piece MoveTo(Square to);
         public abstract List<Square> GetAttacks(Board board);
         public abstract override string ToString();
